Rebase migration

+++ Service/GroupMembershipManagement/Models/AdaptiveCards/ThresholdNotificationCardData.cs
@@ -10,8 +10,8 @@ namespace Models.AdaptiveCards
         public string GroupName { get; set; }
         public int ChangeQuantityForAdditions { get; set; }
         public int ChangeQuantityForRemovals { get; set; }
-        public int ChangePercentageForAdditions { get; set; }
-        public int ChangePercentageForRemovals { get; set; }
+        public double ChangePercentageForAdditions { get; set; }
+        public double ChangePercentageForRemovals { get; set; }
         public int ThresholdPercentageForAdditions { get; set; }
         public int ThresholdPercentageForRemovals { get; set; }
         public string NotificationId { get; set; }
